pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ald-01"
version = "1.0.0"
description = "ALD-01: Advanced Local Desktop Intelligence — Your Personal AI Agent"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ALD-01 Team"},
]
keywords = ["ai", "agent", "assistant", "local", "ollama", "openai", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
voice = [
    "edge-tts>=6.1.0",
    "pyttsx3>=2.90",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0",
    "ruff>=0.2.0",
]

[project.scripts]
ald-01 = "ald01.cli:main"
ald01 = "ald01.cli:main"

[project.urls]
Homepage = "https://github.com/ald-01/ald-01"
Repository = "https://github.com/ald-01/ald-01"
Issues = "https://github.com/ald-01/ald-01/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ald01 = ["dashboard/static/*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.black]
target-version = ["py310"]
line-length = 100
